Fix build.
There is a compilation issue which arise because reglinterface only
exposes a module type.
When index.re is built into cmi, cmx, cma files, the cmx/cma contain
all of the reason code of index.re and reglinterface (which contains no
code) and reglnative. The cmi contains index.rei interface but nothing
for its deps. So when we compile native.re, the module which depends on
index.re, we hit a compilation issue: reglinterface.gl.t isn't defined
This is because the interface file needed for native.re to compile,
but ocamldep doesn't show a direct dependency of native.re on
reglinterface. Ocamldep did its job.

The hack used to fix it was to add a compile time flag to always include
reglinterface.